proc malloc expand /shrink

+++ system/basic/libs/libc/ewoksys/src/proc.c
@@ -67,8 +67,8 @@ inline uint32_t proc_get_uuid(int32_t pid) {
 	return syscall1(SYS_PROC_UUID, pid);
 }
 
-inline void* proc_malloc_expand(uint32_t size) {
-	return (void*)syscall1(SYS_MALLOC_EXPAND, (int32_t)size);
+inline void* proc_malloc_expand(int32_t size) {
+	return (void*)syscall1(SYS_MALLOC_EXPAND, size);
 }
 
 inline void* proc_malloc_free(void) {
